type: Fixes a leak in shared_ptr type interface.
The `TypeInterface` implementation for shared pointers implements the
`dereference` method by calling `initializeStorage` on the pointed type
`TypeInterface`. However, if a null pointer is passed to the
`initializeStorage` as the `storage` parameter, most implementations
allocate a new value and return a pointer to that value.

This behavior has the consequence that when we try to dereference a null
shared pointer, a value is instantiated and returned. Even worse, this
new value is not stored in the shared pointer as it is not possible.

This causes a leak because values returned by the `dereference` function
are not meant to be owned by the caller.

The test that exposed the leak is `TestBind.TestShPtr` in the
`test_messaging` executable.

This patch fixes the issue by changing the behavior of pointer types.
The `dereference` method now must throw a `qi::NullPtrException` (
inheriting `std::runtime_error`) if the pointer is null. It also changes
the type implementations to remove uses of `initializeStorage` in the
`dereference` method.