type: Fixes leak causing by `AnyReferenceCopy`.
The `AnyReferenceCopy` type is used to wrap return values of functions
in `AnyReference` values. It uses an overload of the comma operator to
handle `void` return value types.

In its implementation, it uses an `AssignAnyRef` type that ends up
leaking values in some cases by mishandling references and pointers
values.

This patch fixes this issue by replacing the `AnyReferenceCopy` with a
simple generic function.