Updates supported standard version to C++17. #SW-1427
This patch updates the version of C++ supported by libqi to C++17. Prior
versions of C++ are not supported anymore.

- If the `[[fallthrough]]` attribute is available (introduced in C++17),
  `QI_FALLTHROUGH` expands to it. Before the attribute was available, it
  was, in most cases, expanded to ((void)0).

- `qi::AnyValue` can now contain `std::optional` values. More generally,
  the `std::optional` types are now supported in the type system of
  `qi`.

This patch also fixes build errors occurring when building with C++17.

`std::empty`
============

`std::empty` was added in the standard library in C++17, including
`std::empty(std::initializer_list)`, which conflicts with
`ka::empty(std::initializer_list)` through ADL.

This patch fixes this by adding a condition on the C++ feature on the
declaration of `ka::empty(std::initializer_list)`.

`std::apply`
============
`std::apply` was added in the standard library in C++17. This sometimes
conflicts with `ka::apply` when using ADL (notably because the tuples
and pairs we use for arguments live in the `std` namespace).

This patch fixes this by qualifying calls to `ka::apply` where a
conflict occurs.