messaging: Fixes leak in SDProxy config construction.
A bug in GCC (https://gcc.gnu.org/bugzilla/show_bug.cgi?id=66139, fixed
in GCC12) causes a leak. If the initialization of a struct member
through aggregation (i.e. with the syntax `S { f(), g() }`) throws, the
destructors of the previously constructed struct members are not called.

This causes a leak in our tests on the construction of gateway
configuration. This patch fixes the issue by moving calls to function
that might throw out of the structure initialization.