qi: Removes `Future` and `FutureSync` conversion operator. #SW-1426
This patch removes the conversion operator of `qi::Future<T>` and
`qi::FutureSync<T>` to `const T&`.

These operators have been deprecated for more than 5 years and are
considered harmful. They make an implicit call to a wait operation on
the future result, which blocks the current thread. Blocking without
intention often results in deadlocks in user code.

In order to access the result of a future, user code must now either
create a continuation with `then` or `andThen` or explicitly block
until the result is available with `value` or `wait`.