Fixes noexcept functions advertising SW-2913
This patch fixes advertising `noexcept` functions in libqi, which was broken when compiling the library in C++17.

Since C++17, the noexcept-specification is a part of the function type and may appear as part of any function declarator. This means that template specializations around function types must now include the `noexcept` cases. Unfortunately, the `boost::function_types` library that we used in libqi does not. A replacement called `boost::callable_traits` however does, but where the `function_types` library uses `boost::mpl` heavily, the `boost::callable_traits` instead use standard types such as `std::tuple`.

This patch therefore ports the code using `boost::function_types` to `boost::callable_traits` and does the necessary changes to adapt the meta-programming code. Since most of our uses of `boost::mpl` was because of `function_types`, uses of this library have also been removed.

Furthermore, this patch also fixes some of libqi code that added specializations on function types, which became incomplete (because not handling the `noexcept` specifier cases) in C++17.

Because of these changes, this patch also has to remove or rewrite a lot of hacks and magic that was done around function types. These added a lot of complexity for unknown benefits.