Fixes `ssl::ConfigBase` build error on Clang 12+.
Recent versions of the `clang` compiler generate a build error because
of the protected destructor of the `qi::ssl::ConfigBase` structure:

```
qi/session.hpp:87:58: error: temporary of type 'qi::ssl::ConfigBase' has protected destructor
    Session(bool enforceAuthentication, Config config = {});
```

The issue seems that in C++17, aggregate rules have been widened in the
way that they accept non-virtual base classes, which means that the `{}`
now invokes the aggregate constructor. However, the compiler
rejects aggregates with base classes that have protected destructors.

A defect exists for this issue:
https://www.open-std.org/jtc1/sc22/wg21/docs/cwg_active.html#2244.

The solution here is to either not use the aggregate constructor or to
make the base class destructor public. The protected destructor was
only there to prevent slicing of the objects, but it seems to be a
non-existent problem, and although probably incorrect, slicing is safe.
Therefore the base class protected destructor is now public.