Ensures `ServiceDirectoryProxy` destruction before creation error is returned. #SW-1331
In our gateways, if an error occurs during the creation of the service directory proxy, the
application is stopped and the process is exited, which stops all event loops. It also causes the
destruction of the proxy which in turn calls the destructor of sessions, which waits (blocks) for
sockets disconnection.

However, the socket code does not handle disconnection properly when the process event loops are
stopping, which causes the future returned from the disconnect method to sometimes (due to race
conditions) never finish and consequently causes the destructor of session to block forever. In
turn, the destructor of proxy, that is being executed from a thread of the event loop, blocks, which
blocks the closing of the event loop itself. The process therefore becomes deadlocked and never
exits.

This patch ensures that the destructor of the proxy is executed before the creation error is
returned to the caller, and therefore before the event loops are closed, preventing the deadlock.