messaging: Uses helper function to construct bound object `SocketBinding`. #SW-4667
It occurs that sometimes, objects are bound multiple times to the same socket.
It is unclear at the moment if this happens because of some race condition in
the handling of sockets in our server implementation.

The simplest solution to this problem is to detect these cases by having a
`SocketBinding::make` static function that returns an optional of
`SocketBinding` to the new binding only if it does not exist yet. Callers of
this function may then decide how to react, but usually if a binding already
exists, failure to create the same one is simply ignored.