Update Boost to v1.83.
Boost v1.83 is the version by default shipped by Yocto Nanbield.

Boost.Timer now uses `BOOST_TIMER_ENABLE_DEPRECATED` to enable
deprecated headers. `BOOST_ALLOW_DEPRECATED_HEADERS` is kept for
compatibility with older versions.

The Boost.Unordered library has received a major update, and that has
impacts on `boost::unordered_map`. Some of our tests of
`qi::TraceAnalyzer` incorrectly expected the results of a map traversal
to be fixed when it is arbitrary and depends on the implementation. This
patch fixes that test so that it does not rely on any order.

This patch also adds a missing standard include in `ka/uri/uri_fwd.hpp`
that would cause build errors in some cases.