Fix deprecation warning from Boost.Filesystem.
The `boost::filesystem::symlink_option` is deprecated since Boost 1.72
and should be replaced with `boost::filesystem::directory_options`.
Our only usage of it was to pass
`boost::filesystem::symlink_option::none` to the constructor of
`boost::filesystem::recursive_directory_iterator`, which is the default
with the overload that does not take a
`symlink_option`/`directory_options`, so we prefer using that overload
instead.